+++ Sources/FluidAudio/ASR/AsrTranscription.swift
@@ -28,8 +28,8 @@ extension AsrManager {
 
         if audioSamples.count <= 240_000 {
             let originalLength = audioSamples.count
-            let paddedAudio = padAudioIfNeeded(audioSamples, targetLength: 240_000)
-            let (tokens, _timestamps, encoderSequenceLength) = try await executeMLInferenceWithTimings(
+            let paddedAudio: [Float] = padAudioIfNeeded(audioSamples, targetLength: 240_000)
+            let (tokens, timestamps, encoderSequenceLength) = try await executeMLInferenceWithTimings(
                 paddedAudio,
                 originalLength: originalLength,
                 enableDebug: config.enableDebug,
@@ -38,6 +38,7 @@ extension AsrManager {
 
             let result = processTranscriptionResult(
                 tokenIds: tokens,
+                timestamps: timestamps,
                 encoderSequenceLength: encoderSequenceLength,
                 audioSamples: audioSamples,
                 processingTime: Date().timeIntervalSince(startTime)
@@ -61,23 +62,6 @@ extension AsrManager {
         // ChunkProcessor now uses the passed-in decoder state for continuity
         let processor = ChunkProcessor(audioSamples: audioSamples, enableDebug: config.enableDebug)
         return try await processor.process(using: self, decoderState: &decoderState, startTime: startTime)
-    }
-
-    /// Deprecated: use executeMLInferenceWithTimings and ignore timestamps if not needed
-    @available(*, deprecated, message: "Use executeMLInferenceWithTimings to also retrieve emission timestamps")
-    internal func executeMLInference(
-        _ paddedAudio: [Float],
-        originalLength: Int? = nil,
-        enableDebug: Bool = false,
-        decoderState: inout TdtDecoderState
-    ) async throws -> (tokenIds: [Int], encoderSequenceLength: Int) {
-        let (tokens, _, encLen) = try await executeMLInferenceWithTimings(
-            paddedAudio,
-            originalLength: originalLength,
-            enableDebug: enableDebug,
-            decoderState: &decoderState
-        )
-        return (tokens, encLen)
     }
 
     /// Execute ML inference and return tokens with emission timestamps (encoder frame indices)
@@ -136,6 +120,7 @@ extension AsrManager {
 
     internal func processTranscriptionResult(
         tokenIds: [Int],
+        timestamps: [Int] = [],
         encoderSequenceLength: Int,
         audioSamples: [Float],
         processingTime: TimeInterval,
@@ -144,6 +129,12 @@ extension AsrManager {
 
         let (text, finalTimings) = convertTokensWithExistingTimings(tokenIds, timings: tokenTimings)
         let duration = TimeInterval(audioSamples.count) / TimeInterval(config.sampleRate)
+
+        // Convert timestamps to TokenTiming objects if provided
+        let timingsFromTimestamps = createTokenTimings(from: tokenIds, timestamps: timestamps)
+
+        // Use existing timings if provided, otherwise use timings from timestamps
+        let resultTimings = tokenTimings.isEmpty ? timingsFromTimestamps : finalTimings
 
         if text.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty && duration > 1.0 {
             logger.warning(
@@ -156,13 +147,46 @@ extension AsrManager {
             confidence: 1.0,
             duration: duration,
             processingTime: processingTime,
-            tokenTimings: finalTimings
+            tokenTimings: resultTimings
         )
     }
 
     internal func padAudioIfNeeded(_ audioSamples: [Float], targetLength: Int) -> [Float] {
         guard audioSamples.count < targetLength else { return audioSamples }
         return audioSamples + Array(repeating: 0, count: targetLength - audioSamples.count)
+    }
+
+    /// Convert frame timestamps to TokenTiming objects
+    private func createTokenTimings(from tokenIds: [Int], timestamps: [Int]) -> [TokenTiming] {
+        guard !tokenIds.isEmpty && !timestamps.isEmpty && tokenIds.count == timestamps.count else {
+            return []
+        }
+
+        var timings: [TokenTiming] = []
+
+        for i in 0..<tokenIds.count {
+            let tokenId = tokenIds[i]
+            let frameIndex = timestamps[i]
+
+            // Convert encoder frame index to time (approximate: 80ms per frame)
+            let startTime = TimeInterval(frameIndex) * 0.08
+            let endTime = startTime + 0.08  // Approximate token duration
+
+            // Get token text from vocabulary if available
+            let tokenText = vocabulary[tokenId] ?? "token_\(tokenId)"
+
+            let timing = TokenTiming(
+                token: tokenText,
+                tokenId: tokenId,
+                startTime: startTime,
+                endTime: endTime,
+                confidence: 1.0  // Default confidence
+            )
+
+            timings.append(timing)
+        }
+
+        return timings
     }
 
     /// Slice encoder output to remove left context frames (following NeMo approach)
@@ -196,194 +220,4 @@ extension AsrManager {
         return slicedArray
     }
 
-}
-
-private struct ChunkProcessor {
-    let audioSamples: [Float]
-    let enableDebug: Bool
-
-    // 10 + 2 + 2 seconds context at 16kHz
-    private let sampleRate: Int = 16000
-    private let centerSeconds: Double = 10.0
-    private let leftContextSeconds: Double = 2.0
-    private let rightContextSeconds: Double = 2.0
-
-    private var centerSamples: Int { Int(centerSeconds * Double(sampleRate)) }
-    private var leftContextSamples: Int { Int(leftContextSeconds * Double(sampleRate)) }
-    private var rightContextSamples: Int { Int(rightContextSeconds * Double(sampleRate)) }
-    private var maxModelSamples: Int { 240_000 }  // 15 seconds window capacity
-
-    func process(
-        using manager: AsrManager, decoderState: inout TdtDecoderState, startTime: Date
-    ) async throws -> ASRResult {
-        var allTokens: [Int] = []
-        let audioLength = Double(audioSamples.count) / Double(sampleRate)
-
-        var centerStart = 0
-        var segmentIndex = 0
-        var lastProcessedFrame = 0  // Track the last frame processed by previous chunk
-
-        while centerStart < audioSamples.count {
-            let chunkStartTime = Date()
-            let (windowTokens, maxFrame) = try await processWindowWithTokens(
-                centerStart: centerStart,
-                segmentIndex: segmentIndex,
-                lastProcessedFrame: lastProcessedFrame,
-                using: manager,
-                decoderState: &decoderState
-            )
-            let chunkDuration = Date().timeIntervalSince(chunkStartTime)
-
-            // Update last processed frame for next chunk
-            if maxFrame > 0 {
-                lastProcessedFrame = maxFrame
-            }
-
-            // For chunks after the first, check for and remove duplicated token sequences
-            if segmentIndex > 0 && !allTokens.isEmpty && !windowTokens.isEmpty {
-                let deduplicatedTokens = removeDuplicateSequence(previous: allTokens, current: windowTokens)
-                allTokens.append(contentsOf: deduplicatedTokens)
-
-                if enableDebug && deduplicatedTokens.count != windowTokens.count {
-                    print(
-                        "CHUNK \(segmentIndex): removed \(windowTokens.count - deduplicatedTokens.count) duplicate tokens"
-                    )
-                }
-            } else {
-                allTokens.append(contentsOf: windowTokens)
-            }
-
-            if enableDebug {
-                // Debug: Convert tokens to text for this chunk to see what was produced
-                let (chunkText, _) = manager.convertTokensWithExistingTimings(windowTokens, timings: [])
-                print("CHUNK \(segmentIndex) tokens: \(windowTokens)")
-                print("CHUNK \(segmentIndex) text: '\(chunkText)'")
-                print("ALL tokens so far: \(allTokens)")
-                let (totalText, _) = manager.convertTokensWithExistingTimings(allTokens, timings: [])
-                print("TOTAL text so far: '\(totalText)'")
-                print(
-                    "CHUNK \(segmentIndex): duration=\(String(format: "%.3f", chunkDuration))s, tokens=\(windowTokens.count), total_tokens=\(allTokens.count)"
-                )
-            }
-
-            centerStart += centerSamples
-            segmentIndex += 1
-        }
-
-        let (text, _) = manager.convertTokensWithExistingTimings(allTokens, timings: [])
-
-        return ASRResult(
-            text: text,
-            confidence: 1.0,
-            duration: audioLength,
-            processingTime: Date().timeIntervalSince(startTime),
-            tokenTimings: nil
-        )
-    }
-
-    private func processWindowWithTokens(
-        centerStart: Int,
-        segmentIndex: Int,
-        lastProcessedFrame: Int,
-        using manager: AsrManager,
-        decoderState: inout TdtDecoderState
-    ) async throws -> (tokens: [Int], maxFrame: Int) {
-        // Compute window bounds in samples: [leftStart, rightEnd)
-        let leftStart = max(0, centerStart - leftContextSamples)
-        let centerEnd = min(audioSamples.count, centerStart + centerSamples)
-        let rightEnd = min(audioSamples.count, centerEnd + rightContextSamples)
-
-        // If nothing to process, return empty
-        if leftStart >= rightEnd { return ([], 0) }
-
-        let chunkSamples = Array(audioSamples[leftStart..<rightEnd])
-        let chunkAudioDuration = Double(chunkSamples.count) / Double(sampleRate)
-
-        // Pad to model capacity (15s) if needed; keep track of actual chunk length
-        let paddedChunk = manager.padAudioIfNeeded(chunkSamples, targetLength: maxModelSamples)
-
-        // Calculate encoder frame offset based on where previous chunk ended
-        let startFrameOffset: Int
-        if segmentIndex == 0 {
-            // First chunk: process all frames
-            startFrameOffset = 0
-        } else {
-            // Subsequent chunks: use fixed offset calculation but track last processed frame for filtering
-            let exactEncoderFrameRate = 12.6
-            let leftContextFrames = Int(round(leftContextSeconds * exactEncoderFrameRate))
-            let fixedOffset = leftContextFrames + 2  // Original calculation: ~29
-
-            // Use the fixed offset for consistency, rely on decoder filtering to prevent repetition
-            startFrameOffset = fixedOffset
-
-            print(
-                "CHUNK \(segmentIndex): lastProcessedFrame=\(lastProcessedFrame), fixedOffset=\(fixedOffset), startFrameOffset=\(startFrameOffset)"
-            )
-        }
-
-        print(
-            "CHUNK \(segmentIndex): processing \(String(format: "%.2f", chunkAudioDuration))s audio, leftStart=\(leftStart), centerStart=\(centerStart), rightEnd=\(rightEnd)"
-        )
-        print(
-            "CHUNK \(segmentIndex): startFrameOffset=\(startFrameOffset), skipping \(startFrameOffset) encoder frames"
-        )
-
-        let (tokens, timestamps, encLen) = try await manager.executeMLInferenceWithTimings(
-            paddedChunk,
-            originalLength: chunkSamples.count,
-            enableDebug: false,
-            decoderState: &decoderState,
-            startFrameOffset: startFrameOffset,
-            lastProcessedFrame: lastProcessedFrame
-        )
-
-        if tokens.isEmpty || encLen == 0 {
-            return ([], 0)
-        }
-
-        // Take all tokens from decoder (it already processed only the relevant frames)
-        let filteredTokens = tokens
-        let maxFrame = timestamps.max() ?? 0
-
-        print(
-            "CHUNK \(segmentIndex): audio_duration=\(String(format: "%.2f", chunkAudioDuration))s, startFrameOffset=\(startFrameOffset), taking all \(tokens.count) tokens"
-        )
-
-        return (filteredTokens, maxFrame)
-    }
-
-    /// Remove duplicate token sequences from the beginning of current chunk that match the end of previous chunk
-    private func removeDuplicateSequence(previous: [Int], current: [Int]) -> [Int] {
-        // Look for subsequence matches within reasonable bounds
-        let maxSearchLength = min(15, previous.count)  // Look at last 15 tokens of previous
-        let maxMatchLength = min(12, current.count)  // Look at first 12 tokens of current
-
-        // Ensure we have at least 3 tokens to check for duplication
-        guard maxSearchLength >= 3 && maxMatchLength >= 3 else {
-            return current
-        }
-
-        // Try different overlap lengths, prioritizing longer matches
-        for overlapLength in (3...min(maxSearchLength, maxMatchLength)).reversed() {
-            // Look for this overlap length anywhere in the tail of previous
-            for startIndex in max(0, previous.count - maxSearchLength)..<(previous.count - overlapLength + 1) {
-                let previousSubsequence = Array(previous[startIndex..<startIndex + overlapLength])
-
-                // Check multiple positions in current chunk, not just the beginning
-                for currentStartIndex in 0..<min(5, current.count - overlapLength + 1) {
-                    let currentSubsequence = Array(current[currentStartIndex..<currentStartIndex + overlapLength])
-
-                    if previousSubsequence == currentSubsequence {
-                        print(
-                            "Found duplicate sequence of length \(overlapLength) at position \(currentStartIndex): \(previousSubsequence)"
-                        )
-                        return Array(current.dropFirst(currentStartIndex + overlapLength))
-                    }
-                }
-            }
-        }
-
-        // No duplication found, return original tokens
-        return current
-    }
 }
